build get all future events - need to build the view
