+++ Controllers/DepartmentController.cs
@@ -1,321 +1,321 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.Data.SqlClient;
-using Microsoft.Extensions.Configuration;
-using StudentExercisesMVC.Models;
-using StudentExercisesMVC.Models.ViewModels;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Threading.Tasks;
+//using Microsoft.AspNetCore.Http;
+//using Microsoft.AspNetCore.Mvc;
+//using Microsoft.AspNetCore.Mvc.Rendering;
+//using Microsoft.Data.SqlClient;
+//using Microsoft.Extensions.Configuration;
+//using StudentExercisesMVC.Models;
+//using StudentExercisesMVC.Models.ViewModels;
 
-namespace StudentExercisesMVC.Controllers
-{
-    public class InstructorsController : Controller
+//namespace StudentExercisesMVC.Controllers
+//{
+//    public class InstructorsController : Controller
 
-    {
-        private readonly IConfiguration _config;
+//    {
+//        private readonly IConfiguration _config;
 
-        public InstructorsController(IConfiguration config)
-        {
-            _config = config;
-        }
+//        public InstructorsController(IConfiguration config)
+//        {
+//            _config = config;
+//        }
 
-        public SqlConnection Connection
-        {
-            get
-            {
-                return new SqlConnection(_config.GetConnectionString("DefaultConnection"));
-            }
-        }
-        // GET: Instructors
-        public ActionResult Index()
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    cmd.CommandText = @"SELECT i.Id, i.FirstName, i.LastName, i.SlackHandle, i.Specialty, i.CohortId, c.Name
-                                      FROM Instructor i LEFT JOIN COHORT c 
-                                      ON i.CohortId = c.Id";
+//        public SqlConnection Connection
+//        {
+//            get
+//            {
+//                return new SqlConnection(_config.GetConnectionString("DefaultConnection"));
+//            }
+//        }
+//        // GET: Instructors
+//        public ActionResult Index()
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    cmd.CommandText = @"SELECT i.Id, i.FirstName, i.LastName, i.SlackHandle, i.Specialty, i.CohortId, c.Name
+//                                      FROM Instructor i LEFT JOIN COHORT c 
+//                                      ON i.CohortId = c.Id";
 
-                    SqlDataReader reader = cmd.ExecuteReader();
+//                    SqlDataReader reader = cmd.ExecuteReader();
 
-                    List<Instructor> instructors = new List<Instructor>();
-                    while (reader.Read())
-                    {
-                        Instructor instructor = new Instructor
-                        {
-                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
-                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
-                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
-                            SlackHandle = reader.GetString(reader.GetOrdinal("SlackHandle")),
-                            Specialty = reader.GetString(reader.GetOrdinal("Specialty")),
-                            CohortId = reader.GetInt32(reader.GetOrdinal("CohortId")),
-                            Cohort = new Cohort()
-                            {
-                                Name = reader.GetString(reader.GetOrdinal("Name"))
-                            }
+//                    List<Instructor> instructors = new List<Instructor>();
+//                    while (reader.Read())
+//                    {
+//                        Instructor instructor = new Instructor
+//                        {
+//                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
+//                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
+//                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
+//                            SlackHandle = reader.GetString(reader.GetOrdinal("SlackHandle")),
+//                            Specialty = reader.GetString(reader.GetOrdinal("Specialty")),
+//                            CohortId = reader.GetInt32(reader.GetOrdinal("CohortId")),
+//                            Cohort = new Cohort()
+//                            {
+//                                Name = reader.GetString(reader.GetOrdinal("Name"))
+//                            }
 
-                        };
+//                        };
 
-                        instructors.Add(instructor);
-                    }
+//                        instructors.Add(instructor);
+//                    }
 
-                    reader.Close();
+//                    reader.Close();
 
-                    return View(instructors);
-                }
-            }
+//                    return View(instructors);
+//                }
+//            }
 
-        }
-
-
-        // GET: Students/Details/5
-        public ActionResult Details(int id)
-        {
-            {
-                using (SqlConnection conn = Connection)
-                {
-                    conn.Open();
-                    using (SqlCommand cmd = conn.CreateCommand())
-                    {
-                        cmd.CommandText = "SELECT Id, FirstName, LastName, CohortId, SlackHandle, Specialty FROM Instructor WHERE Id = @id";
-
-                        cmd.Parameters.Add(new SqlParameter("@id", id));
-
-                        var reader = cmd.ExecuteReader();
-                        Instructor instructor = null;
-
-                        if (reader.Read())
-                        {
-                            instructor = new Instructor()
-                            {
-                                Id = reader.GetInt32(reader.GetOrdinal("Id")),
-                                FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
-                                LastName = reader.GetString(reader.GetOrdinal("LastName")),
-                                SlackHandle = reader.GetString(reader.GetOrdinal("SlackHandle")),
-                                Specialty = reader.GetString(reader.GetOrdinal("Specialty")),
-                                CohortId = reader.GetInt32(reader.GetOrdinal("CohortId"))
-                            };
-
-                        }
-                        reader.Close();
-                        return View(instructor);
-                    }
-                }
-            }
-        }
-
-        // GET: Students/Create
-        public ActionResult Create()
-        {
-            var cohortOptions = GetCohortOptions();
-            var viewModel = new InstructorEditViewmodel()
-            {
-                CohortOptions = cohortOptions
-
-            };
-            return View(viewModel);
-        }
-
-        // POST: Instructors/Create
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Create(InstructorEditViewmodel instructor)
-        {
-            try
-            {
-                using (SqlConnection conn = Connection)
-                {
-                    conn.Open();
-                    using (SqlCommand cmd = conn.CreateCommand())
-                    {
-                        cmd.CommandText = @"INSERT INTO Instructor (FirstName, LastName, SlackHandle, Specialty, CohortId)
-                                            OUTPUT INSERTED.Id
-                                            VALUES (@firstName, @lastName, @slackHandle, @specialty, @cohortId)";
-
-                        cmd.Parameters.Add(new SqlParameter("@firstName", instructor.FirstName));
-                        cmd.Parameters.Add(new SqlParameter("@lastName", instructor.LastName));
-                        cmd.Parameters.Add(new SqlParameter("@slackHandle", instructor.SlackHandle));
-                        cmd.Parameters.Add(new SqlParameter("@specialty", instructor.Specialty));
-                        cmd.Parameters.Add(new SqlParameter("@cohortId", instructor.CohortId));
-
-                        var id = (int)cmd.ExecuteScalar();
-                        instructor.Id = id;
-                        // TODO: Add insert logic here
-
-                        return RedirectToAction(nameof(Index));
-                    }
-                }
-            }
-            catch
-            {
-                return View();
-            }
-        }
-
-        // GET: Instructors/Edit/5
-        public ActionResult Edit(int id)
-        {
-            var instructor = GetInstructorById(id);
-            var cohortOptions = GetCohortOptions();
-            var viewModel = new InstructorEditViewmodel()
-            {
-                Id = instructor.Id,
-                FirstName = instructor.FirstName,
-                LastName = instructor.LastName,
-                CohortId = instructor.CohortId,
-                SlackHandle = instructor.SlackHandle,
-                Specialty = instructor.Specialty,
-                CohortOptions = cohortOptions
+//        }
 
 
-            };
-            return View(viewModel);
-        }
+//        // GET: Students/Details/5
+//        public ActionResult Details(int id)
+//        {
+//            {
+//                using (SqlConnection conn = Connection)
+//                {
+//                    conn.Open();
+//                    using (SqlCommand cmd = conn.CreateCommand())
+//                    {
+//                        cmd.CommandText = "SELECT Id, FirstName, LastName, CohortId, SlackHandle, Specialty FROM Instructor WHERE Id = @id";
 
-        // POST: Instructors/Edit/5
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit(int id, InstructorEditViewmodel instructor)
-        {
-            try
-            {
-                using (SqlConnection conn = Connection)
-                {
-                    conn.Open();
-                    using (SqlCommand cmd = conn.CreateCommand())
-                    {
-                        cmd.CommandText = @"UPDATE Instructor 
-                                            SET FirstName = @firstName, 
-                                                LastName = @lastName, 
-                                                SlackHandle = @slackHandle, 
-                                                Specialty = @specialty, 
-                                                CohortId = @cohortId
-                                            WHERE Id = @id";
+//                        cmd.Parameters.Add(new SqlParameter("@id", id));
 
-                        cmd.Parameters.Add(new SqlParameter("@firstName", instructor.FirstName));
-                        cmd.Parameters.Add(new SqlParameter("@lastName", instructor.LastName));
-                        cmd.Parameters.Add(new SqlParameter("@slackHandle", instructor.SlackHandle));
-                        cmd.Parameters.Add(new SqlParameter("@specialty", instructor.Specialty));
-                        cmd.Parameters.Add(new SqlParameter("@cohortId", instructor.CohortId));
-                        cmd.Parameters.Add(new SqlParameter("@id", id));
+//                        var reader = cmd.ExecuteReader();
+//                        Instructor instructor = null;
 
-                        var rowsAffected = cmd.ExecuteNonQuery();
+//                        if (reader.Read())
+//                        {
+//                            instructor = new Instructor()
+//                            {
+//                                Id = reader.GetInt32(reader.GetOrdinal("Id")),
+//                                FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
+//                                LastName = reader.GetString(reader.GetOrdinal("LastName")),
+//                                SlackHandle = reader.GetString(reader.GetOrdinal("SlackHandle")),
+//                                Specialty = reader.GetString(reader.GetOrdinal("Specialty")),
+//                                CohortId = reader.GetInt32(reader.GetOrdinal("CohortId"))
+//                            };
 
-                        if (rowsAffected < 1)
-                        {
-                            return NotFound();
-                        }
-                    }
-                }
+//                        }
+//                        reader.Close();
+//                        return View(instructor);
+//                    }
+//                }
+//            }
+//        }
 
-                return RedirectToAction(nameof(Index));
-            }
-            catch
-            {
-                return View();
-            }
-        }
+//        // GET: Students/Create
+//        public ActionResult Create()
+//        {
+//            var cohortOptions = GetCohortOptions();
+//            var viewModel = new InstructorEditViewmodel()
+//            {
+//                CohortOptions = cohortOptions
 
-        // GET: Instructors/Delete/5
-        public ActionResult Delete(int id)
-        {
-            var instructor = GetInstructorById(id);
-            return View(instructor);
-        }
+//            };
+//            return View(viewModel);
+//        }
+
+//        // POST: Instructors/Create
+//        [HttpPost]
+//        [ValidateAntiForgeryToken]
+//        public ActionResult Create(InstructorEditViewmodel instructor)
+//        {
+//            try
+//            {
+//                using (SqlConnection conn = Connection)
+//                {
+//                    conn.Open();
+//                    using (SqlCommand cmd = conn.CreateCommand())
+//                    {
+//                        cmd.CommandText = @"INSERT INTO Instructor (FirstName, LastName, SlackHandle, Specialty, CohortId)
+//                                            OUTPUT INSERTED.Id
+//                                            VALUES (@firstName, @lastName, @slackHandle, @specialty, @cohortId)";
+
+//                        cmd.Parameters.Add(new SqlParameter("@firstName", instructor.FirstName));
+//                        cmd.Parameters.Add(new SqlParameter("@lastName", instructor.LastName));
+//                        cmd.Parameters.Add(new SqlParameter("@slackHandle", instructor.SlackHandle));
+//                        cmd.Parameters.Add(new SqlParameter("@specialty", instructor.Specialty));
+//                        cmd.Parameters.Add(new SqlParameter("@cohortId", instructor.CohortId));
+
+//                        var id = (int)cmd.ExecuteScalar();
+//                        instructor.Id = id;
+//                        // TODO: Add insert logic here
+
+//                        return RedirectToAction(nameof(Index));
+//                    }
+//                }
+//            }
+//            catch
+//            {
+//                return View();
+//            }
+//        }
+
+//        // GET: Instructors/Edit/5
+//        public ActionResult Edit(int id)
+//        {
+//            var instructor = GetInstructorById(id);
+//            var cohortOptions = GetCohortOptions();
+//            var viewModel = new InstructorEditViewmodel()
+//            {
+//                Id = instructor.Id,
+//                FirstName = instructor.FirstName,
+//                LastName = instructor.LastName,
+//                CohortId = instructor.CohortId,
+//                SlackHandle = instructor.SlackHandle,
+//                Specialty = instructor.Specialty,
+//                CohortOptions = cohortOptions
 
 
-        // POST: Instructors/Delete/5
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Delete(int id, Instructor instructor)
-        {
-            try
-            {
-                using (SqlConnection conn = Connection)
-                {
-                    conn.Open();
-                    using (SqlCommand cmd = conn.CreateCommand())
-                    {
-                        cmd.CommandText = "DELETE FROM Instructor WHERE Id = @id";
-                        cmd.Parameters.Add(new SqlParameter("@id", id));
+//            };
+//            return View(viewModel);
+//        }
 
-                        cmd.ExecuteNonQuery();
-                    }
-                }
+//        // POST: Instructors/Edit/5
+//        [HttpPost]
+//        [ValidateAntiForgeryToken]
+//        public ActionResult Edit(int id, InstructorEditViewmodel instructor)
+//        {
+//            try
+//            {
+//                using (SqlConnection conn = Connection)
+//                {
+//                    conn.Open();
+//                    using (SqlCommand cmd = conn.CreateCommand())
+//                    {
+//                        cmd.CommandText = @"UPDATE Instructor 
+//                                            SET FirstName = @firstName, 
+//                                                LastName = @lastName, 
+//                                                SlackHandle = @slackHandle, 
+//                                                Specialty = @specialty, 
+//                                                CohortId = @cohortId
+//                                            WHERE Id = @id";
 
-                return RedirectToAction(nameof(Index));
-            }
-            catch (Exception ex)
-            {
-                return View();
-            }
-        }
-        private List<SelectListItem> GetCohortOptions()
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    cmd.CommandText = "SELECT Id, Name FROM Cohort";
+//                        cmd.Parameters.Add(new SqlParameter("@firstName", instructor.FirstName));
+//                        cmd.Parameters.Add(new SqlParameter("@lastName", instructor.LastName));
+//                        cmd.Parameters.Add(new SqlParameter("@slackHandle", instructor.SlackHandle));
+//                        cmd.Parameters.Add(new SqlParameter("@specialty", instructor.Specialty));
+//                        cmd.Parameters.Add(new SqlParameter("@cohortId", instructor.CohortId));
+//                        cmd.Parameters.Add(new SqlParameter("@id", id));
+
+//                        var rowsAffected = cmd.ExecuteNonQuery();
+
+//                        if (rowsAffected < 1)
+//                        {
+//                            return NotFound();
+//                        }
+//                    }
+//                }
+
+//                return RedirectToAction(nameof(Index));
+//            }
+//            catch
+//            {
+//                return View();
+//            }
+//        }
+
+//        // GET: Instructors/Delete/5
+//        public ActionResult Delete(int id)
+//        {
+//            var instructor = GetInstructorById(id);
+//            return View(instructor);
+//        }
+
+
+//        // POST: Instructors/Delete/5
+//        [HttpPost]
+//        [ValidateAntiForgeryToken]
+//        public ActionResult Delete(int id, Instructor instructor)
+//        {
+//            try
+//            {
+//                using (SqlConnection conn = Connection)
+//                {
+//                    conn.Open();
+//                    using (SqlCommand cmd = conn.CreateCommand())
+//                    {
+//                        cmd.CommandText = "DELETE FROM Instructor WHERE Id = @id";
+//                        cmd.Parameters.Add(new SqlParameter("@id", id));
+
+//                        cmd.ExecuteNonQuery();
+//                    }
+//                }
+
+//                return RedirectToAction(nameof(Index));
+//            }
+//            catch (Exception ex)
+//            {
+//                return View();
+//            }
+//        }
+//        private List<SelectListItem> GetCohortOptions()
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    cmd.CommandText = "SELECT Id, Name FROM Cohort";
 
 
 
-                    var reader = cmd.ExecuteReader();
-                    var options = new List<SelectListItem>();
+//                    var reader = cmd.ExecuteReader();
+//                    var options = new List<SelectListItem>();
 
-                    while (reader.Read())
-                    {
-                        var option = new SelectListItem()
-                        {
-                            Text = reader.GetString(reader.GetOrdinal("Name")),
-                            Value = reader.GetInt32(reader.GetOrdinal("Id")).ToString()
-                        };
-                        options.Add(option);
+//                    while (reader.Read())
+//                    {
+//                        var option = new SelectListItem()
+//                        {
+//                            Text = reader.GetString(reader.GetOrdinal("Name")),
+//                            Value = reader.GetInt32(reader.GetOrdinal("Id")).ToString()
+//                        };
+//                        options.Add(option);
 
-                    }
-                    reader.Close();
-                    return options;
-                }
-            }
-        }
+//                    }
+//                    reader.Close();
+//                    return options;
+//                }
+//            }
+//        }
 
-        private Instructor GetInstructorById(int id)
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    cmd.CommandText = "SELECT Id, FirstName, LastName, CohortId, SlackHandle, Specialty FROM Instructor WHERE Id = @id";
+//        private Instructor GetInstructorById(int id)
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    cmd.CommandText = "SELECT Id, FirstName, LastName, CohortId, SlackHandle, Specialty FROM Instructor WHERE Id = @id";
 
-                    cmd.Parameters.Add(new SqlParameter("@id", id));
+//                    cmd.Parameters.Add(new SqlParameter("@id", id));
 
-                    var reader = cmd.ExecuteReader();
-                    Instructor instructor = null;
+//                    var reader = cmd.ExecuteReader();
+//                    Instructor instructor = null;
 
-                    if (reader.Read())
-                    {
-                        instructor = new Instructor()
-                        {
-                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
-                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
-                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
-                            SlackHandle = reader.GetString(reader.GetOrdinal("SlackHandle")),
-                            Specialty = reader.GetString(reader.GetOrdinal("Specialty")),
-                            CohortId = reader.GetInt32(reader.GetOrdinal("CohortId"))
-                        };
+//                    if (reader.Read())
+//                    {
+//                        instructor = new Instructor()
+//                        {
+//                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
+//                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
+//                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
+//                            SlackHandle = reader.GetString(reader.GetOrdinal("SlackHandle")),
+//                            Specialty = reader.GetString(reader.GetOrdinal("Specialty")),
+//                            CohortId = reader.GetInt32(reader.GetOrdinal("CohortId"))
+//                        };
 
-                    }
-                    reader.Close();
-                    return instructor;
-                }
-            }
-        }
-    }
-}
+//                    }
+//                    reader.Close();
+//                    return instructor;
+//                }
+//            }
+//        }
+//    }
+//}
