+++ appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost\\SQLEXPRESS;Database=StudentExercises;Trusted_Connection=True"
+    "DefaultConnection": "Server=localhost\\SQLEXPRESS;Database=BangazonWorkforce;Trusted_Connection=True"
   }
 }
